+++ src/builtins/builtins.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "internal.h"
+#include "internal.hh"
 
 namespace rego
 {
