+++ src/helpers.cc
@@ -1,4 +1,4 @@
-#include "internal.h"
+#include "internal.hh"
 #include "version.h"
 
 #ifdef _WIN32
