+++ src/passes/structure.cc
@@ -1,4 +1,4 @@
-#include "internal.h"
+#include "internal.hh"
 
 namespace
 {
